requirements.txt
flask
psycopg2-binary
sqlalchemy
alembic
python-dotenv
gunicorn
